% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{fmpv3_path}
\alias{fmpv3_path}
\title{FMP Cloud V3 Path}
\usage{
fmpv3_path(path = "earning_calendar", ...)
}
\arguments{
\item{path}{Path added to Fmp cloud host}

\item{...}{Query params to GET request}
}
\value{
Result of GET request
}
\description{
Get data fro FMP cloud using API v3.
}
\examples{
fmpv3_path("earning_calendar", from = Sys.date() - 5, to = Sys.date(), apikey = Sys.getenv("APIKEY-FMPCLOUD"))
}
