% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{fmpv_path}
\alias{fmpv_path}
\title{FMP Cloud V3 Path}
\usage{
fmpv_path(path = "earning_calendar", v = "v3", ...)
}
\arguments{
\item{path}{Path added to Fmp cloud host}

\item{v}{API version}

\item{...}{Query params to GET request}
}
\value{
Result of GET request
}
\description{
Get data fro FMP cloud using API v3.
}
