% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_utils.R
\name{get_all_blob_files}
\alias{get_all_blob_files}
\title{Get All Blob Files}
\usage{
get_all_blob_files(container = "transcripts")
}
\arguments{
\item{container}{Blob container}
}
\value{
Result of GET request
}
\description{
Get all blob files from specifed container.
}
