% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_utils.R
\name{get_all_blob_files_content}
\alias{get_all_blob_files_content}
\title{Import Files From Container}
\usage{
get_all_blob_files_content(
  save_file,
  container = "transcripts",
  refresh_data_old = 30,
  ...
)
}
\arguments{
\item{save_file}{Path to local file. This is where merged file is saved and path to load data from if it is already saved}

\item{container}{Blob container}

\item{refresh_data_old}{Look at refresh_data_old in get_blob_file}

\item{...}{other arguments to get_blob_file function.}
}
\value{
Result of GET request
}
\description{
Import all blob files from container
}
