% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_utils.R
\name{get_blob_file}
\alias{get_blob_file}
\title{Get File from blob}
\usage{
get_blob_file(file, container = "fundamentals")
}
\arguments{
\item{file}{File to import (without extensions)}

\item{container}{Blob container}
}
\value{
Result of GET request
}
\description{
Get files from Azure blob storage. Files can be in csv and rds format.
}
