% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_prices.R
\name{get_daily_prices}
\alias{get_daily_prices}
\title{Get Daily prices}
\usage{
get_daily_prices(
  symbols,
  start_date = Sys.Date() - 30,
  end_date = Sys.Date(),
  blob_file = NA,
  blob_cont = "fundamentals"
)
}
\arguments{
\item{symbols}{Start date}

\item{start_date}{End dateevents}

\item{end_date}{End date}

\item{blob_file}{Name of blob file. If NA, the file wan't be saved on Azure Store blob.}

\item{blob_cont}{Azure storage blob container}
}
\value{
Result of GET request
}
\description{
Get daily prices for specified stocks
}
