% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_earning_announcements.R
\name{get_earning_announcements}
\alias{get_earning_announcements}
\title{Get Earning Announcement Data}
\usage{
get_earning_announcements(
  start_date = Sys.Date() - 1,
  end_date = Sys.Date(),
  apikey = Sys.getenv("APIKEY-FMPCLOUD"),
  local_file = NA
)
}
\arguments{
\item{start_date}{Start date}

\item{end_date}{End date}

\item{apikey}{Fmp cloud API key}

\item{local_file}{PPth to local file. Use only if you want to save file localy.}
}
\value{
Result of GET request
}
\description{
Get earnings announcements data from FMP cloud
}
\examples{
fmpv3_path("earning_calendar", from = Sys.date() - 5, to = Sys.date(), apikey = Sys.getenv("APIKEY-FMPCLOUD"))
}
