% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_earning_surprises.R
\name{get_earning_surprises}
\alias{get_earning_surprises}
\title{Get Earning Surprises Data}
\usage{
get_earning_surprises(symbols, blob_file = NA)
}
\arguments{
\item{symbols}{Stock symbols}

\item{blob_file}{Name of blob file}
}
\value{
Data table with earning surprises
}
\description{
Get earnings surprises data from FMP cloud
}
