% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_investingcom_earnings_calendar_bulk.R
\name{get_investingcom_earnings_calendar}
\alias{get_investingcom_earnings_calendar}
\title{Get Earnings Calendar from investing.com}
\usage{
get_investingcom_earnings_calendar(date_from, date_to)
}
\arguments{
\item{date_from}{Start date}

\item{date_to}{End date}
}
\value{
Result of GET request
}
\description{
Get Earnings Calendar from investing.com for specified dates
}
