% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_investingcom_earnings_calendar_bulk.R
\name{get_investingcom_earnings_calendar_bulk}
\alias{get_investingcom_earnings_calendar_bulk}
\title{Get Complete Earnings Calendar from investing.com}
\usage{
get_investingcom_earnings_calendar_bulk(start_date, blob_file = NA)
}
\arguments{
\item{start_date}{Start date}

\item{blob_file}{Name of blob file}
}
\value{
Result of GET request
}
\description{
Get complete earnings calendar from investing.com.
}
