% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risk_quarks.R
\name{get_risk_quarks}
\alias{get_risk_quarks}
\title{Get Complete Earnings Calendar from yahoo}
\usage{
get_risk_quarks(
  prices,
  p = c(0.975),
  win_size = 200,
  model = c("EWMA", "GARCH"),
  method = c("plain", "age", "vwhs", "fhs"),
  forecast_length = 100,
  container = "risk-quarks-day"
)
}
\arguments{
\item{start_date}{Start date}

\item{end_date}{End date}

\item{blob_file}{Blob file}
}
\value{
Data frame with yahoo earnings calnedar data
}
\description{
Get complete earnings calendar from yahoo.
}
