% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tick_data.R
\name{get_tick_data}
\alias{get_tick_data}
\title{Get Tick Data from FMP cloud}
\usage{
get_tick_data(symbol, date)
}
\arguments{
\item{symbol}{Start date}

\item{date}{End date}
}
\value{
Data frame with columns timestamp (SIP Unix Timestamp), ask (Ask price), bid (Bid price), bsize (Bid size - lots), asize (Ask size - lots)
}
\description{
Get and save tick data from FMP cloud
}
