% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transcripts.R
\name{get_transcripts}
\alias{get_transcripts}
\title{Get Transcripts}
\usage{
get_transcripts(
  symbols,
  years = 1992:2021,
  blob_cont = "transcripts",
  overwrite_existing_files = FALSE
)
}
\arguments{
\item{symbols}{stock symbols}

\item{years}{years to extract data from.}

\item{blob_cont}{blob container}

\item{overwrite_existing_files}{should existing files be overwritten}
}
\value{
Result of GET request
}
\description{
Get Earning Call Transcript from FMP cloud.
}
