% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yahoo_calendar_all.R
\name{get_yahoo_calendar}
\alias{get_yahoo_calendar}
\title{Get Earnings Calendar from yahoo}
\usage{
get_yahoo_calendar(date)
}
\arguments{
\item{date}{Date}
}
\value{
Data frame with yahoo earnings calnedar data
}
\description{
Get earnings calendar from yahoo for specific date.
}
