% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_utils.R
\name{save_blob_files}
\alias{save_blob_files}
\title{Save files to blob}
\usage{
save_blob_files(object, file_name, container = "fundamentals")
}
\arguments{
\item{object}{Object to save}

\item{file_name}{File to import (without extensions)}

\item{container}{Blob container}
}
\value{
Result of GET request
}
\description{
Save files bob. It automaticly saves files as csv and rds objects
}
