% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_earning_announcements.R
\name{update_earning_announcements}
\alias{update_earning_announcements}
\title{Update Earning Announcement Data}
\usage{
update_earning_announcements(blob_file = "earnings-calendar.rds")
}
\arguments{
\item{blob_file}{Blob file to save updated data}
}
\value{
Result of GET request
}
\description{
Update earnings announcements data from FMP cloud
}
