% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_utils.R
\name{get_blob_file}
\alias{get_blob_file}
\title{Get File from blob}
\usage{
get_blob_file(
  blob_file,
  container = "fundamentals",
  save_file = file.path(getwd(), file),
  refresh_data_old = 3,
  ...
)
}
\arguments{
\item{blob_file}{File to import from Azure store container}

\item{container}{Blob container}

\item{save_file}{File name to save object localy}

\item{refresh_data_old}{Number of days after which new, refreshed data would be donwloaded fom container.}

\item{...}{Other arguments to storage_download function from AzureStore pakcage}
}
\value{
Result of GET request
}
\description{
Get files from Azure blob storage. Files can be in csv and rds format.
}
