% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_earning_announcements.R
\name{get_earning_announcements}
\alias{get_earning_announcements}
\title{Get Earning Announcement Data}
\usage{
get_earning_announcements(
  start_date = Sys.Date() - 10,
  end_date = Sys.Date(),
  blob_file = NA
)
}
\arguments{
\item{start_date}{Start date}

\item{end_date}{End date}

\item{blob_file}{Name of blob file}
}
\value{
Result of GET request
}
\description{
Get earnings announcements data from FMP cloud
}
