% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_earning_announcements.R
\name{get_earning_announcements}
\alias{get_earning_announcements}
\title{Get Earning Announcement Data}
\usage{
get_earning_announcements(
  start_date = Sys.Date() - 1,
  end_date = Sys.Date(),
  local_file = "D:/fundamental_data/earnings_announcement/earning-calendar",
  blob_file = "earning-calendar"
)
}
\arguments{
\item{start_date}{Start date}

\item{end_date}{End date}

\item{local_file}{Path to local file. Use only if you want to save file localy.}

\item{blob_file}{Name of blob file}
}
\value{
Result of GET request
}
\description{
Get earnings announcements data from FMP cloud
}
